+++ src/app/admin/roles/MainRoles.tsx
@@ -1,0 +1,278 @@
+'use client'
+import Typography from '@mui/material/Typography'
+import { ReactNode, useEffect, useState } from 'react'
+import { RolCRUDType } from '@/types/roles'
+import { useSession } from '@/hooks/useSession'
+import { useAuth } from '@/context/AuthProvider'
+import { CasbinTypes } from '@/types/utils/casbin'
+import { Button, Grid, useMediaQuery, useTheme } from '@mui/material'
+import {
+  delay,
+  InterpreteMensajes,
+  siteName,
+  titleCase,
+} from '@/utils/utilidades'
+import { usePathname } from 'next/navigation'
+import { Constantes } from '@/config'
+import { CriterioOrdenType } from '@/types/ordenTypes'
+import { AlertDialog } from '@/components/modales/AlertDialog'
+import { imprimir } from '@/utils/imprimir'
+import { ordenFiltrado } from '@/utils/orden'
+import { toast } from 'sonner'
+
+import { BotonBuscar } from '@/components/botones/BotonBuscar'
+import CustomMensajeEstado from '@/components/estados/CustomMensajeEstado'
+import { IconoTooltip } from '@/components/botones/IconoTooltip'
+import { BotonOrdenar } from '@/components/botones/BotonOrdenar'
+import { IconoBoton } from '@/components/botones/IconoBoton'
+import { Paginacion } from '@/components/datatable/Paginacion'
+import { CustomDialog } from '@/components/modales/CustomDialog'
+import { CustomDataTable } from '@/components/datatable/CustomDataTable'
+
+import { VistaModalRol } from './ModalRol'
+import { FiltroRol } from './FiltroRol'
+
+export default function RolesPage() {
+  const [rolesData, setRolesData] = useState<RolCRUDType[]>([])
+  const [loading, setLoading] = useState<boolean>(true)
+
+  const [errorRolData, setErrorRolData] = useState<any>()
+
+  const [modalRol, setModalRol] = useState(false)
+
+  const [rolEdicion, setRolEdicion] = useState<RolCRUDType | undefined>()
+
+  // Variables de páginado
+  const [limite, setLimite] = useState<number>(10)
+  const [pagina, setPagina] = useState<number>(1)
+  const [total, setTotal] = useState<number>(0)
+
+  //filtros
+
+  const [filtroRol, setFiltroRol] = useState<string>('')
+  const [mostrarFiltroRol, setMostrarFiltroRol] = useState(false)
+
+  const { sesionPeticion } = useSession()
+
+  const theme = useTheme()
+  const xs = useMediaQuery(theme.breakpoints.only('xs'))
+
+  const [mostrarAlertaEstadoRol, setMostrarAlertaEstadoRol] = useState(false)
+
+  const editarEstadoRolModal = (rol: RolCRUDType) => {
+    setRolEdicion(rol) // para mostrar datos de Rol en la alerta
+    setMostrarAlertaEstadoRol(true) // para mostrar alerta de Roles
+  }
+
+  const cancelarAlertaEstadoRol = async () => {
+    setMostrarAlertaEstadoRol(false)
+    await delay(500)
+    setRolEdicion(undefined)
+  }
+
+  const aceptarAlertaEstadoRol = async () => {
+    setMostrarAlertaEstadoRol(false)
+    if (rolEdicion) {
+      await cambiarEstadoRolPeticion(rolEdicion)
+    }
+    setRolEdicion(undefined)
+  }
+
+  // router para conocer la ruta actual
+  const pathname = usePathname()
+
+  /// Criterios de orden
+  const [ordenCriterios, setOrdenCriterios] = useState<
+    Array<CriterioOrdenType>
+  >([
+    { campo: 'rol', nombre: 'Rol', ordenar: true },
+    { campo: 'nombre', nombre: 'Nombre', ordenar: true },
+    { campo: 'estado', nombre: 'Estado', ordenar: true },
+    { campo: 'acciones', nombre: 'Acciones' },
+  ])
+
+  const acciones: Array<ReactNode> = [
+    <BotonBuscar
+      id={'accionFiltrarRolToggle'}
+      key={'accionFiltrarRolToggle'}
+      seleccionado={mostrarFiltroRol}
+      cambiar={setMostrarFiltroRol}
+    />,
+    xs && (
+      <BotonOrdenar
+        id={'ordenarRoles'}
+        key={`ordenarRoles`}
+        label={'Ordenar roles'}
+        criterios={ordenCriterios}
+        cambioCriterios={setOrdenCriterios}
+      />
+    ),
+    <IconoTooltip
+      id={'actualizarRol'}
+      titulo={'Actualizar'}
+      key={`accionActualizarRol`}
+      accion={async () => {
+        await obtenerRolesPeticion()
+      }}
+      icono={'refresh'}
+      name={'Actualizar lista de roles'}
+    />,
+    permisos.create && (
+      <IconoBoton
+        id={'agregarRol'}
+        key={'agregarRol'}
+        texto={'Agregar'}
+        variante={xs ? 'icono' : 'boton'}
+        icono={'add_circle_outline'}
+        descripcion={'Agregar rol'}
+        accion={() => {
+          agregarRolModal()
+        }}
+      />
+    ),
+  ]
+
+  const cambiarEstadoRolPeticion = async (rol: RolCRUDType) => {
+    try {
+      setLoading(true)
+      const respuesta = await sesionPeticion({
+        url: `${Constantes.baseUrl}/autorizacion/roles/${rol.id}/${
+          rol.estado == 'ACTIVO' ? 'inactivacion' : 'activacion'
+        }`,
+        tipo: 'patch',
+      })
+      imprimir(`respuesta inactivar rol: ${respuesta}`)
+      toast.success(InterpreteMensajes(respuesta))
+      await obtenerRolesPeticion()
+    } catch (e) {
+      imprimir(`Error al inactivar rol`, e)
+      toast.error(InterpreteMensajes(e))
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  const obtenerRolesPeticion = async () => {
+    try {
+      setLoading(true)
+
+      const respuesta = await sesionPeticion({
+        url: `${Constantes.baseUrl}/autorizacion/roles/todos`,
+        params: {
+          pagina: pagina,
+          limite: limite,
+          ...(filtroRol.length == 0 ? {} : { filtro: filtroRol }),
+          ...(ordenFiltrado(ordenCriterios).length == 0
+            ? {}
+            : {
+                orden: ordenFiltrado(ordenCriterios).join(','),
+              }),
+        },
+      })
+      setRolesData(respuesta.datos?.filas)
+      setTotal(respuesta.datos?.total)
+      setErrorRolData(null)
+    } catch (e) {
+      imprimir(`Error al obtener Roles`, e)
+      setErrorRolData(e)
+      toast.error(InterpreteMensajes(e))
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  const agregarRolModal = () => {
+    setRolEdicion(undefined)
+    setModalRol(true)
+  }
+  const editarRolModal = (Rol: RolCRUDType) => {
+    setRolEdicion(Rol)
+    setModalRol(true)
+  }
+
+  const cerrarModalRol = async () => {
+    setModalRol(false)
+    await delay(500)
+    setRolEdicion(undefined)
+  }
+
+  useEffect(() => {
+    obtenerRolesPeticion().finally(() => {})
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [
+    pagina,
+    limite,
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    JSON.stringify(ordenCriterios),
+    filtroRol,
+  ])
+
+  useEffect(() => {
+    if (!mostrarFiltroRol) {
+      setFiltroRol('')
+    }
+  }, [mostrarFiltroRol])
+
+  const paginacion = (
+    <Paginacion
+      pagina={pagina}
+      limite={limite}
+      total={total}
+      cambioPagina={setPagina}
+      cambioLimite={setLimite}
+    />
+  )
+
+  return (
+    <>
+      <title>{`Roles - ${siteName()}`}</title>
+      <AlertDialog
+        isOpen={mostrarAlertaEstadoRol}
+        titulo={'Alerta'}
+        texto={`¿Está seguro de ${
+          rolEdicion?.estado == 'ACTIVO' ? 'inactivar' : 'activar'
+        } a ${titleCase(rolEdicion?.nombre ?? '')} ?`}
+      >
+        <Button onClick={cancelarAlertaEstadoRol}>Cancelar</Button>
+        <Button onClick={aceptarAlertaEstadoRol}>Aceptar</Button>
+      </AlertDialog>
+      <CustomDialog
+        isOpen={modalRol}
+        handleClose={cerrarModalRol}
+        title={rolEdicion ? 'Editar rol' : 'Nuevo rol'}
+      >
+        <VistaModalRol
+          rol={rolEdicion}
+          accionCorrecta={() => {
+            cerrarModalRol().finally()
+            obtenerRolesPeticion().finally()
+          }}
+          accionCancelar={cerrarModalRol}
+        />
+      </CustomDialog>
+      <CustomDataTable
+        titulo={'Roles'}
+        error={!!errorRolData}
+        cargando={loading}
+        acciones={acciones}
+        columnas={ordenCriterios}
+        cambioOrdenCriterios={setOrdenCriterios}
+        paginacion={paginacion}
+        contenidoTabla={contenidoTabla}
+        filtros={
+          mostrarFiltroRol && (
+            <FiltroRol
+              filtroRol={filtroRol}
+              accionCorrecta={(filtros) => {
+                setPagina(1)
+                setLimite(10)
+                setFiltroRol(filtros.rol)
+              }}
+              accionCerrar={() => {}}
+            />
+          )
+        }
+      />
+    </>
+  )
+}
