+++ src/app/admin/roles/page.tsx
@@ -1,36 +1,36 @@
 'use client'
-import Typography from '@mui/material/Typography'
-import { ReactNode, useEffect, useState } from 'react'
-import { RolCRUDType } from '@/types/roles'
-import { useSession } from '@/hooks/useSession'
+
 import { useAuth } from '@/context/AuthProvider'
+import { useSession } from '@/hooks/useSession'
+import { RolCRUDType } from '@/types/roles'
+import { RUTAS, RutasType, obtenerPermisos } from '@/types/temporalTypes'
 import { CasbinTypes } from '@/types/utils/casbin'
-import { Button, Grid, useMediaQuery, useTheme } from '@mui/material'
-import { delay, InterpreteMensajes } from '@/utils/utilidades'
-// import { delay, InterpreteMensajes, siteName, titleCase } from '@/utils'
-import { usePathname } from 'next/navigation'
-import { Constantes } from '@/config'
-import { CriterioOrdenType } from '@/types/ordenTypes'
-import { AlertDialog } from '@/components/modales/AlertDialog'
 import { imprimir } from '@/utils/imprimir'
-import { ordenFiltrado } from '@/utils/orden'
-import { toast } from 'sonner'
+import { delay, siteName } from '@/utils/utilidades'
+import { useMediaQuery, useTheme } from '@mui/material'
+import { notFound, usePathname } from 'next/navigation'
+import { ReactNode, useState } from 'react'
 
-import { BotonBuscar } from '@/components/botones/BotonBuscar'
-import CustomMensajeEstado from '@/components/estados/CustomMensajeEstado'
-import { IconoTooltip } from '@/components/botones/IconoTooltip'
-import { BotonOrdenar } from '@/components/botones/BotonOrdenar'
-import { IconoBoton } from '@/components/botones/IconoBoton'
-import { Paginacion } from '@/components/datatable/Paginacion'
-import { CustomDialog } from '@/components/modales/CustomDialog'
-import { CustomDataTable } from '@/components/datatable/CustomDataTable'
+const obtenerRutasPermisos = (ruta: string, rol: any): CasbinTypes => {
+  /// Fetch api que devuelva los permisos dado un rol de la ruta que se está enviando Ver SolicitarPermisos interface
+  const existe: RutasType | undefined = RUTAS.find((elem) => elem.ruta === ruta)
 
-import { VistaModalRol } from './ModalRol'
-import { FiltroRol } from './FiltroRol'
+  if (existe === undefined) notFound()
+  const permisos = obtenerPermisos(existe.permisos)
+  return permisos
+}
 
-export default function RolesPage() {
+const RolesPage = () => {
+  const pathname = usePathname()
+  const { rolUsuario } = useAuth()
+  imprimir('--------------------------------')
+  imprimir(rolUsuario && rolUsuario.rol)
+  imprimir('--------------------------------')
+  const rolUsuarioNombre = rolUsuario?.rol
+
+  const permisos: CasbinTypes = obtenerRutasPermisos(pathname, rolUsuarioNombre)
+
   const [rolesData, setRolesData] = useState<RolCRUDType[]>([])
-  const [loading, setLoading] = useState<boolean>(true)
 
   const [errorRolData, setErrorRolData] = useState<any>()
 
@@ -47,17 +47,6 @@ export default function RolesPage() {
 
   const [filtroRol, setFiltroRol] = useState<string>('')
   const [mostrarFiltroRol, setMostrarFiltroRol] = useState(false)
-
-  const { sesionPeticion } = useSession()
-  const { permisoUsuario } = useAuth()
-
-  // Permisos para acciones
-  const [permisos, setPermisos] = useState<CasbinTypes>({
-    read: false,
-    create: false,
-    update: false,
-    delete: false,
-  })
 
   const theme = useTheme()
   const xs = useMediaQuery(theme.breakpoints.only('xs'))
@@ -82,19 +71,6 @@ export default function RolesPage() {
     }
     setRolEdicion(undefined)
   }
-
-  // router para conocer la ruta actual
-  const pathname = usePathname()
-
-  /// Criterios de orden
-  const [ordenCriterios, setOrdenCriterios] = useState<
-    Array<CriterioOrdenType>
-  >([
-    { campo: 'rol', nombre: 'Rol', ordenar: true },
-    { campo: 'nombre', nombre: 'Nombre', ordenar: true },
-    { campo: 'estado', nombre: 'Estado', ordenar: true },
-    { campo: 'acciones', nombre: 'Acciones' },
-  ])
 
   const contenidoTabla: Array<Array<ReactNode>> = rolesData.map(
     (rolData, indexRol) => [
@@ -149,47 +125,6 @@ export default function RolesPage() {
     ]
   )
 
-  const acciones: Array<ReactNode> = [
-    <BotonBuscar
-      id={'accionFiltrarRolToggle'}
-      key={'accionFiltrarRolToggle'}
-      seleccionado={mostrarFiltroRol}
-      cambiar={setMostrarFiltroRol}
-    />,
-    xs && (
-      <BotonOrdenar
-        id={'ordenarRoles'}
-        key={`ordenarRoles`}
-        label={'Ordenar roles'}
-        criterios={ordenCriterios}
-        cambioCriterios={setOrdenCriterios}
-      />
-    ),
-    <IconoTooltip
-      id={'actualizarRol'}
-      titulo={'Actualizar'}
-      key={`accionActualizarRol`}
-      accion={async () => {
-        await obtenerRolesPeticion()
-      }}
-      icono={'refresh'}
-      name={'Actualizar lista de roles'}
-    />,
-    permisos.create && (
-      <IconoBoton
-        id={'agregarRol'}
-        key={'agregarRol'}
-        texto={'Agregar'}
-        variante={xs ? 'icono' : 'boton'}
-        icono={'add_circle_outline'}
-        descripcion={'Agregar rol'}
-        accion={() => {
-          agregarRolModal()
-        }}
-      />
-    ),
-  ]
-
   const cambiarEstadoRolPeticion = async (rol: RolCRUDType) => {
     try {
       setLoading(true)
@@ -210,136 +145,11 @@ export default function RolesPage() {
     }
   }
 
-  const obtenerRolesPeticion = async () => {
-    try {
-      setLoading(true)
-
-      const respuesta = await sesionPeticion({
-        url: `${Constantes.baseUrl}/autorizacion/roles/todos`,
-        params: {
-          pagina: pagina,
-          limite: limite,
-          ...(filtroRol.length == 0 ? {} : { filtro: filtroRol }),
-          ...(ordenFiltrado(ordenCriterios).length == 0
-            ? {}
-            : {
-                orden: ordenFiltrado(ordenCriterios).join(','),
-              }),
-        },
-      })
-      setRolesData(respuesta.datos?.filas)
-      setTotal(respuesta.datos?.total)
-      setErrorRolData(null)
-    } catch (e) {
-      imprimir(`Error al obtener Roles`, e)
-      setErrorRolData(e)
-      toast.error(InterpreteMensajes(e))
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const agregarRolModal = () => {
-    setRolEdicion(undefined)
-    setModalRol(true)
-  }
-  const editarRolModal = (Rol: RolCRUDType) => {
-    setRolEdicion(Rol)
-    setModalRol(true)
-  }
-
-  const cerrarModalRol = async () => {
-    setModalRol(false)
-    await delay(500)
-    setRolEdicion(undefined)
-  }
-
-  async function definirPermisos() {
-    setPermisos(await permisoUsuario(pathname))
-  }
-
-  useEffect(() => {
-    definirPermisos().finally()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  useEffect(() => {
-    obtenerRolesPeticion().finally(() => {})
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [
-    pagina,
-    limite,
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    JSON.stringify(ordenCriterios),
-    filtroRol,
-  ])
-
-  useEffect(() => {
-    if (!mostrarFiltroRol) {
-      setFiltroRol('')
-    }
-  }, [mostrarFiltroRol])
-
-  const paginacion = (
-    <Paginacion
-      pagina={pagina}
-      limite={limite}
-      total={total}
-      cambioPagina={setPagina}
-      cambioLimite={setLimite}
-    />
-  )
-
   return (
     <>
       <title>{`Roles - ${siteName()}`}</title>
-      <AlertDialog
-        isOpen={mostrarAlertaEstadoRol}
-        titulo={'Alerta'}
-        texto={`¿Está seguro de ${
-          rolEdicion?.estado == 'ACTIVO' ? 'inactivar' : 'activar'
-        } a ${titleCase(rolEdicion?.nombre ?? '')} ?`}
-      >
-        <Button onClick={cancelarAlertaEstadoRol}>Cancelar</Button>
-        <Button onClick={aceptarAlertaEstadoRol}>Aceptar</Button>
-      </AlertDialog>
-      <CustomDialog
-        isOpen={modalRol}
-        handleClose={cerrarModalRol}
-        title={rolEdicion ? 'Editar rol' : 'Nuevo rol'}
-      >
-        <VistaModalRol
-          rol={rolEdicion}
-          accionCorrecta={() => {
-            cerrarModalRol().finally()
-            obtenerRolesPeticion().finally()
-          }}
-          accionCancelar={cerrarModalRol}
-        />
-      </CustomDialog>
-      <CustomDataTable
-        titulo={'Roles'}
-        error={!!errorRolData}
-        cargando={loading}
-        acciones={acciones}
-        columnas={ordenCriterios}
-        cambioOrdenCriterios={setOrdenCriterios}
-        paginacion={paginacion}
-        contenidoTabla={contenidoTabla}
-        filtros={
-          mostrarFiltroRol && (
-            <FiltroRol
-              filtroRol={filtroRol}
-              accionCorrecta={(filtros) => {
-                setPagina(1)
-                setLimite(10)
-                setFiltroRol(filtros.rol)
-              }}
-              accionCerrar={() => {}}
-            />
-          )
-        }
-      />
     </>
   )
 }
+
+export default RolesPage
